requirements.txt
flask
africastalking
python-dotenv

